requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
loguru==0.5.3
Markdown==3.3.3
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
